添加api
作品/专辑 ＋ 喜欢/收藏 + 添加/取消
作品/专辑 ＋ 喜欢/收藏 ＋ 是否

+++ Tumcca/Tumcca/src/com/pineapple/mobilecraft/tumcca/activity/AlbumWorkListActivity.java
@@ -21,7 +21,7 @@ import java.util.List;
 public class AlbumWorkListActivity extends FragmentActivity {
 
     private TextView mTvTitle;
-    public static void startActivity(Activity activity, int id, String albumName){
+    public static void startActivity(Activity activity, long id, String albumName){
         Intent intent = new Intent(activity, AlbumWorkListActivity.class);
         intent.putExtra("id", id);
         intent.putExtra("albumName", albumName);
